feat: configs: tidy up configs origin
