--- src/stylistic.ts
@@ -1,3 +1,0 @@
-import pkg from '@stylistic/eslint-plugin';
-
-export const { configs: stylisticConfigs } = pkg;
