--- src/vue.ts
@@ -1,3 +1,0 @@
-import pkg from 'eslint-plugin-vue';
-
-export const { configs: vueConfigs } = pkg;
